+++ backend/src/handlers/poker-tables/PokerTablesJoinHandler.test.ts
