+++ backend/src/handlers/poker-tables/PokerTablesLeaveHandler.test.ts
