+++ backend/src/shared/api/PokerTables/errors/PokerTableDoesNotExistError.ts
@@ -1,0 +1,6 @@
+import { BaseError } from '@Infra/Result';
+export class PokerTableDoesNotExistError extends BaseError {
+  constructor() {
+    super('table_does_not_exist', 'Table does not exist');
+  }
+}
