+++ backend/src/shared/api/PokerTables/errors/SeatTakenError.ts
@@ -1,0 +1,7 @@
+import { BaseError } from '@Infra/Result';
+
+export class SeatTakenError extends BaseError {
+  constructor() {
+    super('seat_taken', 'Seat is taken');
+  }
+}
