+++ backend/src/shared/api/PokerTables/types/PokerTableJoin.ts
@@ -1,0 +1,18 @@
+// Types
+import { BaseOutput } from '../../BaseOutput';
+
+// External
+import Joi from 'joi';
+
+export type PokerTableJoinPayload = {
+  selectedSeatNumber: string;
+  socketId: string;
+};
+
+export interface PokerTableJoinOutput extends BaseOutput {}
+
+// Joi schema
+export const pokerTableJoinSchema = Joi.object<PokerTableJoinPayload>({
+  selectedSeatNumber: Joi.string().required(),
+  socketId: Joi.string().required(),
+});
