+++ backend/src/shared/api/PokerTables/types/PokerTableLeave.ts
@@ -1,0 +1,18 @@
+// Types
+import { BaseOutput } from '../../BaseOutput';
+
+// External
+import Joi from 'joi';
+
+export type PokerTableLeavePayload = {
+  selectedSeatNumber: string;
+  socketId: string;
+};
+
+export interface PokerTableLeaveOutput extends BaseOutput {}
+
+// Joi schema
+export const pokerTableLeaveSchema = Joi.object<PokerTableLeavePayload>({
+  selectedSeatNumber: Joi.string().required(),
+  socketId: Joi.string().required(),
+});
