+++ backend/src/shared/api/Users/errors/UserNameTakenError.ts
@@ -1,0 +1,7 @@
+import { BaseError } from '@Infra/Result';
+
+export class UserNameTakenError extends BaseError {
+  constructor() {
+    super('username_taken', 'Username taken');
+  }
+}
