+++ backend/src/shared/api/Users/errors/UsersCreateError.ts
@@ -1,0 +1,7 @@
+import { BaseError } from '@Infra/Result';
+
+export class UsersCreateError extends BaseError {
+  constructor() {
+    super('user_create_error', 'User not created');
+  }
+}
