+++ backend/src/tests/mocks/dbMocks.ts
@@ -3,7 +3,7 @@ import { RowDataPacket } from 'mysql2/promise';
 
 // Internal
 import { Result, ResultSuccess, ResultError } from '@Infra/Result';
-import { MySqLInstance } from '../../db/MySql';
+import { MySqLInstance } from '../../db/my-sql';
 import { DBInsertError, DBInsertDuplicateError } from '../../db/errors/DBInsertErrors';
 import { DBSelectError } from '../../db/errors/DBSelectErrors';
